rename check files; comment old code
